Replace span interface by smart pointer to const
Heavily improve documentation
Rewrite LiCount to take advantage of builtin exponential integral
Replace 64 bit and 16 bit integers with 32 bit where appropriate for
speed
Re-add Moebius functions to iterators

+++ BitArray.h
@@ -1,6 +1,5 @@
 #pragma once
 
-#include <cstddef>
 #include <cstdint>
 #include <vector>
 
@@ -9,7 +8,7 @@ class BitArray
 {
 private:
     // The underlying storage.
-    std::vector<uint64_t> storage;
+    std::vector<uint32_t> storage;
 
     // The number of bits stored.
     size_t count;
@@ -19,12 +18,15 @@ public:
     BitArray (size_t count, bool defaultValue);
 
     // Gets the bit at a given index.
+    // Out of range indices result in undefined behaviour.
     bool Get (size_t index) const;
 
     // Sets true the bit at a given index.
+    // Out of range indices result in undefined behaviour.
     void Set (size_t index);
 
     // Sets false the bit at a given index.
+    // Out of range indices result in undefined behaviour.
     void Reset (size_t index);
 
     // Returns the number of bits stored.
